Authorization functionality updated

+++ Qtec.AccountManagement/Src/Qtec.AccountManagement.Web/Pages/ChartsOfAccounts/ChartTree.cshtml.cs
@@ -1,19 +1,24 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Qtec.AccountManagement.Application.Services;
 using Qtec.AccountManagement.Domain.Dtos;
-using Serilog;
 
 namespace Qtec.AccountManagement.Web.Pages.ChartsOfAccounts
 {
+    [Authorize]
     public class ChartTreeModel : PageModel
     {
         private readonly AccountManagementService _accountManagementService;
+        private readonly IAuthorizationService _authorizationService;
         private readonly ILogger<ChartTreeModel> _logger;
 
-        public ChartTreeModel(AccountManagementService accountManagementService, ILogger<ChartTreeModel> logger )
+        public ChartTreeModel(AccountManagementService accountManagementService, 
+            IAuthorizationService authorizationService,
+            ILogger<ChartTreeModel> logger )
         {
             _accountManagementService = accountManagementService;
+            _authorizationService = authorizationService;
             _logger = logger;
         }
         public List<AccountDto> TreeView { get; set; } = new();
@@ -33,6 +38,12 @@ namespace Qtec.AccountManagement.Web.Pages.ChartsOfAccounts
         {
             try
             {
+                var result = await _authorizationService.AuthorizeAsync(User, "Admin");
+
+                if (!result.Succeeded)
+                {
+                    return RedirectToPage("/AccessDenied");
+                }
                 var success = await _accountManagementService.DeleteAccountAsync(id);
                 if (success)
                 {
